Move DynamicQuantizeLinear shape inference into a quantize submodule
Move the DynamicQuantizeLinear operator, plus its test, from ops.rs into a
new ops/quantize.rs submodule, matching how this operator is grouped in
src/ops.rs.
